+++ Source/Runtime/YaGE/Graphics/CommandBuffer.h
@@ -185,7 +185,7 @@ public:
                               uint32_t     height,
                               DXGI_FORMAT  srcFormat,
                               const void  *src,
-                              uint32_t     srcRowPitch,
+                              size_t       srcRowPitch,
                               PixelBuffer &dest,
                               uint32_t     mipLevel) -> void;
